Refactor repeats to their own module so commands can start using them
